Description: fix problems IssueNo: https://gitee.com/openharmony/ace_engine_lite/issues/I4Z4J8 Feature or Bugfix: Bugfix Binary Source:No Signed-off-by: liuyuxiang <liuyuxiang7@huawei.com>

+++ frameworks/tools/qt/simulator/jsfwk/targets/simulator/mock/amsthread/ams_thread.h
@@ -39,7 +39,7 @@ struct JSAppRecord {
     ~JSAppRecord()
     {
         if (jsApp != nullptr) {
-            const unsigned long waitTime = 20;
+            const uint32_t waitTime = 20;
             while (jsApp->isRunning()) {
                 QThread::msleep(waitTime);
             }
